+++ src/Model/Usuario.php
@@ -6,12 +6,12 @@ use \RedBeanPHP\R as R;
 
 class Usuario
 {
-/**
- * Obtém dados do usuário
- *
- * @param int $codpes
- * @return bean usuario | array mensagem de erro
- */
+    /**
+     * Obtém dados do usuário
+     *
+     * @param int $codpes
+     * @return bean usuario | array mensagem de erro
+     */
     public static function obter($codpes)
     {
         $usuario = R::findOne('usuario', 'codpes = ?', [$codpes]);
@@ -20,5 +20,10 @@ class Usuario
             exit;
         }
         return $usuario;
+    }
+
+    public static function obterOuCriar($codpes)
+    {
+        return R::findOrCreate('usuario', ['codpes' => $codpes]);
     }
 }